Create MiSTerTrigger.sh
This script checks for a trigger file on two MiSTer devices. If the trigger file exists on a device, the script removes the trigger file and starts the synchronization script. The script then stops running. This script should be placed in the /etc/ directory on your Pi and your MiSTer devices should be configured to allow FTP access.